+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - 2018-11-08
+### Fixed
+- Race condition for adding products on PDP
+
 ## [1.0.2] - 2018-10-24
 ### Added
 - Cache for customer data